feat : 상품 리스트 조회

+++ src/main/java/com/example/backoffice/product/dto/CreateProductResponse.java
@@ -1,8 +1,8 @@
 package com.example.backoffice.product.dto;
 
-import com.example.backoffice.admin.entity.Administrator;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -10,12 +10,13 @@ public class CreateProductResponse {
     private final Long id;
     private final String name;
     private final String category;
-    private final Long price;
+    private final BigDecimal price;
     private final Long stock;
     private final String status;
     private final LocalDateTime createdAt;
 
-    public CreateProductResponse(Long id, String name, String category, Long price, Long stock, String status, LocalDateTime createdAt) {
+
+    public CreateProductResponse(Long id, String name, String category, BigDecimal price, Long stock, String status, LocalDateTime createdAt) {
         this.id = id;
         this.name = name;
         this.category = category;
